fix(mcp): keep text-search fallback usable without node:sqlite
The fallback backend imported splitIntoSections from build-index,
which statically imports node:sqlite. On Node < 22 (and non-Node
runtimes) linking that module threw ERR_UNKNOWN_BUILTIN_MODULE and
crashed the server before it could fall back — defeating the whole
point of the fallback.

Move the pure MDX parsing helpers into a new sqlite-free parse
module. build-index re-exports them and still owns the node:sqlite
index writer; the naive backend now imports only from parse, so its
module graph never touches node:sqlite.

Also make the fallback warning a loud deprecation banner: Node < 22
is no longer supported and the text-search path will be removed.
